+++ Project.toml
@@ -5,6 +5,8 @@ version = "0.2.0"
 
 [deps]
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
+Parameters = "d96e819e-fc66-5662-9728-84c9c7592b0a"
+Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 Transducers = "28d57a85-8fef-5791-bfe6-a80928e7c999"
 
 [compat]
